+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.3.0-rc.45",
+  sdkVersion: "4.3.0-rc.46",
   genVersion: "2.814.0",
-  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.45 2.814.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.46 2.814.0 2.0 @orq-ai/node",
 } as const;